+++ collections/README.md
@@ -510,19 +510,19 @@ console.assert(
 );
 ```
 
-### takeFirstWhile
+### takeWhile
 
 Returns all elements in the given collection until the first element that does
 not match the given predicate.
 
 ```ts
-import { takeFirstWhile } from "https://deno.land/std@$STD_VERSION/collections/mod.ts";
+import { takeWhile } from "https://deno.land/std@$STD_VERSION/collections/mod.ts";
 import { assertEquals } from "https://deno.land/std@$STD_VERSION/testing/asserts.ts";
 
 const arr = [1, 2, 3, 4, 5, 6];
 
 assertEquals(
-  takeFirstWhile(arr, (i) => i !== 4),
+  takeWhile(arr, (i) => i !== 4),
   [1, 2, 3],
 );
 ```
