+++ collections/mod.ts
@@ -26,4 +26,4 @@ export * from "./union.ts";
 export * from "./without_all.ts";
 export * from "./unzip.ts";
 export * from "./zip.ts";
-export * from "./take_first_while.ts";
+export * from "./take_while.ts";
