+++ collections/take_while.ts
@@ -4,18 +4,18 @@
  *
  * Example:
  * ```ts
- * import { takeFirstWhile } from "./take_first_while.ts";
+ * import { takeWhile } from "./take_while.ts";
  * import { assertEquals } from "../testing/asserts.ts";
  *
  * const arr = [1, 2, 3, 4, 5, 6];
  *
  * assertEquals(
- *   takeFirstWhile(arr, (i) => i !== 4),
+ *   takeWhile(arr, (i) => i !== 4),
  *   [1, 2, 3],
  * );
  * ```
  */
-export function takeFirstWhile<T>(
+export function takeWhile<T>(
   array: readonly T[],
   predicate: (el: T) => boolean,
 ): T[] {
